refactor(settings): remove unused components and streamline settings structure
- Deleted `PageHeader` and `Tabs` components to simplify the codebase.
- Removed `dimmed` prop from `SettingRow` to enhance clarity and reduce complexity.
- Consolidated imports in `settingsContent` and `TopLevelSections` for better organization.
- Added new tests for detail panel functionality to ensure robust coverage.
- Cleaned up various test files by removing obsolete tests and adding new ones for updated components.
